+++ src/fft/ifft_mfa_truncate_sqrt2.c
@@ -139,8 +139,8 @@ void ifft_mfa_truncate_sqrt2(mp_limb_t ** ii, mp_size_t n, flint_bitcnt_t w,
    flint_bitcnt_t depth2 = 0;
    flint_bitcnt_t limbs = (w*n)/FLINT_BITS;
 
-   while ((UWORD(1)<<depth) < n2) depth++;
-   while ((UWORD(1)<<depth2) < n1) depth2++;
+   while ((UWORD(1)<<depth) < (ulong) n2) depth++;
+   while ((UWORD(1)<<depth2) < (ulong) n1) depth2++;
 
    /* first half mfa IFFT : n2 rows, n1 cols */
 
@@ -431,8 +431,8 @@ void ifft_mfa_truncate_sqrt2_outer(mp_limb_t ** ii, mp_size_t n, flint_bitcnt_t 
     thread_pool_handle * threads;
     ifft_outer_arg_t * args;
 
-    while ((UWORD(1)<<depth) < n2) depth++;
-    while ((UWORD(1)<<depth2) < n1) depth2++;
+    while ((UWORD(1)<<depth) < (ulong) n2) depth++;
+    while ((UWORD(1)<<depth2) < (ulong) n1) depth2++;
 
 #if FLINT_USES_PTHREAD
     pthread_mutex_init(&mutex, NULL);
